+++ src/App.js
@@ -3,12 +3,14 @@ import Shared from "./page/shared/Shared";
 import Folder from "./page/folder/Folder";
 
 function App(){
-  <BrowserRouter>
-    <Routes>
-      <Route path="/" element={<Shared />}></Route>
-      <Route path="/" element={<Folder />}></Route>
-    </Routes>
-  </BrowserRouter>
+  return(
+    <BrowserRouter>
+      <Routes>
+        <Route path="/" element={<Shared />}></Route>
+        <Route path="/folder" element={<Folder />}></Route>
+      </Routes>
+    </BrowserRouter>
+  )
 }
 
 export default App;
