+++ src/api.js
@@ -1,8 +1,8 @@
-const BASE_URL = "https://bootcamp-api.codeit.kr/api/users/1";
+const BASE_URL = "https://bootcamp-api.codeit.kr/api";
 
 async function getProfile() {
   try{
-    const respones = await fetch(`${BASE_URL}`);
+    const respones = await fetch(`${BASE_URL}/users/1`);
     if(!respones.ok) throw new Error('데이터를 불러오는데 실패했습니다');
     
     const result = await respones.json();
@@ -14,7 +14,7 @@ async function getProfile() {
 
 async function getFolder() {
   try{
-    const respones = await fetch(`${BASE_URL}/links`);
+    const respones = await fetch(`${BASE_URL}/users/1/links`);
     if(!respones.ok) throw new Error('데이터를 불러오는데 실패했습니다');
     
     const result = await respones.json();
@@ -26,7 +26,7 @@ async function getFolder() {
 
 async function getFolderMenu() {
   try{
-    const respones = await fetch(`${BASE_URL}/folders`);
+    const respones = await fetch(`${BASE_URL}/users/1/folders`);
     if(!respones.ok) throw new Error('데이터를 불러오는데 실패했습니다');
     
     const result = await respones.json();
@@ -36,4 +36,29 @@ async function getFolderMenu() {
   }
 }
 
-export { getProfile, getFolder, getFolderMenu }
+async function getProfileAsync() {
+  try{
+    const respones = await fetch(`${BASE_URL}/sample/user`);
+    if(!respones.ok) throw new Error('데이터를 불러오는데 실패했습니다');
+    
+    const result = await respones.json();
+    return result;
+  }catch(error){
+    console.log(error);
+  }
+}
+
+async function getFolderAsync() {
+  try{
+    const respones = await fetch(`${BASE_URL}/sample/folder`);
+    if(!respones.ok) throw new Error('데이터를 불러오는데 실패했습니다');
+    
+    const result = await respones.json();
+    return result;
+  }catch(error){
+    console.log(error);
+  }
+}
+
+
+export { getProfile, getFolder, getFolderMenu, getProfileAsync, getFolderAsync }
