--- src/components/SearchInput.js
@@ -1,13 +1,0 @@
-import '../page/folder/Main.css';
-
-function SearchInput(){
-  return(
-    <div className="main">
-      <div className="search">
-        <input type="search" placeholder="링크를 검색해 보세요."/>
-      </div>
-    </div>
-  )
-}
-
-export default SearchInput;
