+++ src/components/searchBar/SearchBar.js
@@ -1,0 +1,11 @@
+import './SearchBar.css';
+
+function SearchBar(){
+  return(
+    <div className="search">
+      <input type="search" placeholder="링크를 검색해 보세요."/>
+    </div>
+  )
+}
+
+export default SearchBar;
