+++ src/components/tabMenu/TabMenu.css
@@ -1,0 +1,41 @@
+.main .tab-menu-wrap {
+  position: relative;
+}
+
+.main .tab-menu ul {
+  display: flex;
+  gap: 0 0.8rem;
+  list-style-type: none;
+  margin: 4rem 0 0;
+}
+
+.main .tab-menu ul li {
+  line-height: 1.9rem;
+}
+
+.main .tab-menu ul li button {
+  display: block;
+  padding: 0 1.2rem;
+  height: 3.5rem;
+  font-size: 1.6rem;
+  border-radius: 5px;
+  border: 1px solid var(--color-blue);
+  background-color: var(--color-white);
+  cursor: pointer;
+}
+
+.main .tab-menu ul li button.active {
+  color: var(--color-white);
+  background-color: var(--color-blue);
+}
+
+@media screen and (min-width:375px) and (max-width:768px) {
+  .main .tab-menu ul {
+    flex-wrap: wrap;
+    gap: 1.2rem 0.8rem;
+  }
+
+  .main .tab-menu ul li button {
+    font-size: 1.4rem;
+  }
+}
