+++ src/index.js
@@ -1,10 +1,9 @@
 import ReactDOM from "react-dom/client";
 import App from "./App";
-import Folder from "./page/folder/Folder";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
   <>
-    <Folder />
+    <App />
   </>
 );
