+++ src/page/folder/Folder.js
@@ -1,17 +1,88 @@
 import Footer from "../../components/Footer";
-import "../folder/Header.css";
 import "../../components/reset.css";
 import "../../components/root.css";
-import Profile from '../../components/Profile';
-import Data from '../../components/Data';
-import SearchInput from '../../components/SearchInput';
+import { useEffect, useState } from "react";
+import { getProfile, getFolder, getFolderMenu } from "../../api";
+import Header from './Header';
+import Main from './Main';
+
+const BASE_URL = "https://bootcamp-api.codeit.kr/api/users/1";
 
 function Folder() {
+  const [profileImg, setProfileImg] = useState(null);
+  const [profileEmail, setProfileEmail] = useState("");
+  const [cardList, setCardList] = useState([]);
+  const [folderMenu, setFolderMenu] = useState([]);
+  const [menuActive, setMenuActive] = useState('all');
+  const [btnOption, setBtnOption] = useState(false);
+  const [title, setTitle] = useState('');
+
+  const handleLoadProfile = async () => {
+    const { data } = await getProfile();
+    setProfileImg(data[0].image_source);
+    setProfileEmail(data[0].email);
+  };
+
+  const handleLoadFolderMenu = async () => {
+    const { data } = await getFolderMenu();
+    setFolderMenu(data);
+  };
+
+  const handleClick = (item) => {
+    setMenuActive(item.id);
+    setBtnOption(true);
+    setTitle(`${item.name !== '전체' ? item.name : ""}`);
+  }
+
+  const handleMouseOver = (e) => e.currentTarget.classList.add("active");
+
+  const handleMouseOut = (e) => e.currentTarget.classList.remove("active");
+
+  const handleLoadFolder = async (options) => {
+    if(options !== 'all') {
+      const { data } = await getFolderList(options);
+      setCardList(data);
+    }else{
+      const { data } = await getFolder();
+      setCardList(data);
+    }
+  };
+
+  async function getFolderList(menuActive) {
+    try{
+      const respones = await fetch(`${BASE_URL}/links?folderId=${menuActive}`);
+      if(!respones.ok) throw new Error('데이터를 불러오는데 실패했습니다');
+      
+      const result = await respones.json();
+      return result;
+    }catch(error){
+      console.log(error);
+    }
+  }
+
+  useEffect(() => {
+    handleLoadProfile();
+    handleLoadFolderMenu();
+    handleLoadFolder(menuActive);
+  }, [menuActive]);
+
   return (
     <div className="container">
-      <Profile/>
-      <SearchInput />
-      <Data />
+      <Header 
+        profileImg={profileImg}
+        profileEmail={profileEmail}
+      />
+      <Main
+        className="main"
+        links={cardList}
+        menu={folderMenu}
+        onMouseOver={handleMouseOver}
+        onMouseOut={handleMouseOut}
+        menuActive={menuActive}
+        handleClick={handleClick}
+        btnOption={btnOption}
+        title={title}
+      />
       <Footer className="footer"/>
     </div>
   );
