+++ src/page/folder/cardList/CardList.css
@@ -1,0 +1,110 @@
+.main .card ul {
+  list-style-type: none;
+  display: flex;
+  flex-wrap: wrap;
+  gap: 2.5rem 0;
+  justify-content: space-between;
+}
+
+.main .card ul li {
+  width: calc(33.3% - 2rem);
+  box-shadow: 0 5px 25px rgba(0, 0, 0, 0.08);
+  border-radius: 15px;
+}
+
+.main .card ul li a {
+  display: block;
+  overflow: hidden;
+  border: 3px solid transparent;
+  border-radius: 15px;
+}
+
+.main .card ul li.active a {
+  border: 3px solid var(--color-blue);
+}
+
+.main .card ul li .img-box {
+  overflow: hidden; 
+  position: relative;
+  height: 20rem;
+}
+
+.main .card ul li .img-box .star {
+  position: absolute;
+  top: 1.6rem;
+  right: 1.6rem;
+  width: 3.4rem;
+  height: 3.4rem;
+}
+
+.main .card ul li img {
+  display: block;
+  width: 100%;
+  height: 20rem;
+  object-fit: cover;
+}
+
+.main .card ul li.active .img-card {
+  transform: scale(1.3);
+  transition: all 0.5s;
+}
+
+.main .card ul li .card-text {
+  overflow: hidden;
+  min-height: 11.5rem;
+  padding: 1.5rem 2rem;
+}
+
+.main .card ul li.active .card-text {
+  background-color: var(--bg-sky-blue);
+}
+
+.main .card ul li .card-text .created-at {
+  float: left;
+  font-size: 1.3rem;
+  line-height: 1.5rem;
+  color: var(--color-gray);
+}
+
+.main .card ul li .card-text .kabab {
+  display: block;
+  float: right;
+  width: 2.1rem;
+  height: 1.7rem;
+  background: url('../../../assets/btn-kebab.png') no-repeat;
+}
+
+.main .card ul li .card-text .desc {
+  overflow: hidden;
+  display: -webkit-box;
+  -webkit-line-clamp: 2;
+  -webkit-box-orient: vertical;
+  text-overflow: ellipsis;
+  min-height: 3.8rem;
+  margin: 2.1rem 0 1rem 0;
+  font-size: 1.6rem;
+  color: var(--color-black-000);
+  clear: both;
+}
+
+.main .card ul li .card-text .date {
+  font-size: 1.4rem;
+  line-height: 1.6rem;
+  color: var(--color-black-333);
+}
+
+@media screen and (max-width:1124px) {
+  .main .card ul li {
+    width: calc(50% - 2.4rem);
+  }
+}
+
+@media screen and (min-width:375px) and (max-width:768px) {
+  .main .card ul {
+    gap: 2rem 0;
+  }
+
+  .main .card ul li {
+    width: 100%;
+  }
+}
