+++ src/page/shared/Header.js
@@ -1,39 +1,27 @@
-import './Header.css';
-import logo from '../../assets/logo.svg';
+import FolderUser from '../../components/FolderUser/FolderUser'
+import Nav from './nav/Nav';
 
-function Header({ className, profileEmail, profileImg, folderUserProfile, folderUserName, folderName }){
+function SharedHeader({ 
+  className, 
+  profileEmail, 
+  profileImg, 
+  folderName, 
+  folderUserName, 
+  folderUserProfile 
+}){
   return (
     <header className={className}>
-      <nav>
-        <h1 className="logo">
-          <a href="/">
-            <img src={logo} alt="홈으로 연결된 Linkbrary 로고" />
-          </a>
-        </h1>
-        <div className="profile">
-          {profileEmail ? 
-          <a href="/">
-           <img src={profileImg} alt="프로필 이미지" />
-           <p className="text mobile-hide">
-             {profileEmail}
-           </p>
-         </a> :
-          <a href="signin.html" className="cta cta-short">
-            <span>로그인</span>
-          </a>}
-        </div>
-      </nav> 
-      <div className="folder-info">
-        <img className="user-profile" src={folderUserProfile} alt="폴더 사용자 프로필 이미지" />
-        <p className="user-name">
-          {folderUserName}
-        </p>
-        <p className="name">
-          {folderName}
-        </p>
-      </div>
+      <Nav 
+        profileEmail={profileEmail} 
+        profileImg={profileImg}
+      />
+      <FolderUser 
+        folderName={folderName}
+        folderUserName={folderUserName}
+        folderUserProfile={folderUserProfile}
+      />
     </header>
-  )
+  );
 }
 
-export default Header;
+export default SharedHeader;
