+++ src/page/shared/Shared.js
@@ -4,9 +4,7 @@ import Footer from "../../components/Footer";
 import "../../components/reset.css";
 import "../../components/root.css";
 import { useEffect, useState } from "react";
-import useAsync from "../../hooks/useAsync";
-
-const BASE_URL = "https://bootcamp-api.codeit.kr/api/sample";
+import { getProfileAsync, getFolderAsync } from "../../api";
 
 function Shared() {
   const [cardList, setCardList] = useState([]);
@@ -15,36 +13,20 @@ function Shared() {
   const [folderUserProfile, setFolderUserProfile] = useState(null);
   const [folderUserName, setFolderUserName] = useState("");
   const [folderName, setFolderName] = useState("");
-  const [profileIsLoading, profileError, getProfileAsync] = useAsync(
-    `${BASE_URL}/user`
-  );
-  const [folderIsLoading, folderError, getFolderAsync] = useAsync(
-    `${BASE_URL}/folder`
-  );
 
   const handleLoadProfile = async () => {
-    const result = await getProfileAsync();
-    if (!result) return;
-
-    const { email, profileImageSource } = result;
+    const { email, profileImageSource } = await getProfileAsync();
     setProfileImg(profileImageSource);
     setProfileEmail(email);
-
-    return [profileIsLoading, profileError];
   };
 
   const handleLoadFolder = async () => {
-    const result = await getFolderAsync();
-    if (!result) return;
-
-    const { folder } = result;
+    const { folder } = await getFolderAsync();
     const { links } = folder;
     setFolderName(folder.name);
     setFolderUserName(folder?.owner?.name);
     setFolderUserProfile(folder?.owner?.profileImageSource);
     setCardList(links);
-
-    return [folderIsLoading, folderError];
   };
 
   const handleMouseOver = (e) => e.currentTarget.classList.add("active");
