+++ src/page/shared/cardList/CardList.css
@@ -1,35 +1,3 @@
-.main {
-  margin: 4rem 0 10rem 0;
-  padding: 0 19rem;
-}
-
-.main .search {
-  position: relative;
-}
-
-.main .search::before {
-  content: '';
-  display: block;
-  position: absolute;
-  left: 1.7rem;
-  top: 50%;
-  transform: translateY(-50%);
-  width: 1.6rem;
-  height: 1.6rem;
-  background: url('../../assets/ico-search.png') no-repeat;
-}
-
-.main .search input {
-  width: 100%;
-  padding: 1.5rem 4.2rem;
-  outline: none;
-  border: none;
-  border-radius: 1rem;
-  font-size: 1.6rem;
-  color: var(--color-gray);
-  background-color: var(--color-very-light-gray);
-}
-
 .main .card ul {
   list-style-type: none;
   display: flex;
@@ -105,26 +73,12 @@
 }
 
 @media screen and (max-width:1124px) {
-  .main {
-    padding: 0 3.2rem;
-  }
-
   .main .card ul li {
     width: calc(50% - 2.4rem);
   }
 }
 
 @media screen and (min-width:375px) and (max-width:768px) {
-  .main {
-    margin: 2rem 0 6rem 0;
-  }
-
-  .main .search input {
-    padding: 1.3rem 3.8rem;
-    font-size: 1.4rem;
-    line-height: 1.6rem;
-  }
-
   .main .card ul {
     gap: 2rem 0;
     margin: 3.2rem 0 0;
@@ -134,6 +88,3 @@
     width: 100%;
   }
 }
-
-
-
